refactor(fetcher): split fetcher-health module and fix failure nav
Extract js/fetcher/ submodules behind a stable barrel export, route the global
status pill to Connections on auth failures, and stop clearing sticky failed
state when opening the fetcher popover.
